% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_manipulation.R
\name{add_category}
\alias{add_category}
\title{Add a New Category Column to a Dataframe Based on Target Content}
\usage{
add_category(
  df,
  target_column,
  target_content,
  new_column_content = NULL,
  new_column_name = category,
  drop_rows = FALSE
)
}
\arguments{
\item{df}{A dataframe to which the new category column will be added.}

\item{target_column}{The name of the column in the dataframe where the content
matching will occur. This column is used as the basis for determining the new
category column's content.}

\item{target_content}{A character vector specifying the content to match in the
\code{target_column}. Each element represents a pattern to match.}

\item{new_column_content}{(Optional) A character vector specifying the new
content to add to the \code{new_column_name} based on matching the \code{target_content}.
If not provided, the default behavior is to insert NAs.}

\item{new_column_name}{(Optional) The name of the new category column to be added
to the dataframe. Default is 'category'.}

\item{drop_rows}{(Optional) Logical indicating whether rows that match the
\code{target_content} should be dropped. Default is FALSE.}
}
\value{
A dataframe with the added (or modified) category column.
}
\description{
This function adds a new column to a dataframe, populating it based on matching
content in a target column. It can also drop rows that match specified content.
}
\examples{
\dontrun{
tibble(a = c("hi", "here", "hi", "bye", "target2", "other")) \%>\%
  add_category(a, c("here", "target2"),
               c("One", "Two"),
               drop_rows = TRUE)
}

}
