% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate.R
\name{add_index}
\alias{add_index}
\title{Add an index column to a data frame or tibble}
\usage{
add_index(df, col_name = index)
}
\arguments{
\item{df}{A data frame or tibble to which the index will be added.}

\item{col_name}{The name to assign to the new index column. Defaults to 'index'.}
}
\value{
A data frame or tibble with the added index column.
}
\description{
This function adds an index column to a data frame or tibble.
The index starts from 1 and increments by 1 for each row.
If the data is grouped, the index will be relative to the group.
}
\examples{
\dontrun{
# Adding index to an ungrouped tibble
tibble(a = c(3, 3, 3, 2, 2, 3)) \%>\%
  add_index()

# Adding index to a grouped tibble
tibble(a = c(3, 3, 3, 2, 2, 3)) \%>\%
  group_by(a) \%>\%
  add_index(col_name = "group.index")

# Changing the name of the index column
tibble(a = c(3, 3, 3, 2, 2, 3)) \%>\%
  add_index(col_name = "row_id")
  }
}
