% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_temp_column.R
\name{add_temp_column}
\alias{add_temp_column}
\title{Add a Temporary Column Name}
\usage{
add_temp_column(df)
}
\arguments{
\item{df}{A dataframe to which the temporary column name will be added.}
}
\value{
A character string representing the unique temporary column name.
}
\description{
Generates a unique temporary column name for a given dataframe \code{df}.
The function starts with a base name of "temp" and appends a number to it
if "temp" already exists as a column name. The number is incremented until a unique
column name is found.
}
\examples{
\dontrun{
df_example <- tibble(a = 1:3, temp = 4:6, temp1 = 7:9)
add_temp_column(df_example)
}

}
