% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_manipulation.R
\name{apply_columnwise}
\alias{apply_columnwise}
\title{Apply a Function Columnwise to Two List-Columns in a Data Frame}
\usage{
apply_columnwise(df, col1, col2, .f, output_col = "result", ...)
}
\arguments{
\item{df}{A data frame containing the list-columns to be processed.}

\item{col1}{The name of the first list-column (as a symbol or string).}

\item{col2}{The name of the second list-column (as a symbol or string).}

\item{.f}{The function to apply to corresponding columns of the list-columns.
This function should take at least two arguments.}

\item{output_col}{The name of the output list-column (default is "result").}

\item{...}{Additional arguments to pass to the function .f.}
}
\value{
A data frame with an additional list-column containing the results
of applying .f to corresponding columns of col1 and col2.
}
\description{
This function takes a data frame and two list-columns containing tibbles
with identical structures. It applies a specified function to corresponding
columns in these tibbles and returns a new list-column containing the results.
}
\examples{
\dontrun{
# Example usage
example_df <- tibble(
  ID = 1:3,
  events = list(tibble(col1 = 1:3, col2 = 4:6)),
  pt = list(tibble(col1 = 2:4, col2 = 5:7))
)
result <- apply_columnwise(example_df, events, pt, function(x, y) x - y)
}
}
