% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.r
\name{chrom_lengths<-}
\alias{chrom_lengths<-}
\title{Add Chromosome Lengths to Dataframe attributes}
\usage{
chrom_lengths(.data) <- value
}
\arguments{
\item{.data}{Dataframe containing genomic range data.}

\item{chrom_lengths}{Dataframe of chromosome lengths (with columns
\code{chrom} and \code{length}.}
}
\description{
Add Chromosome Lengths to Dataframe attributes
}
