% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.r
\name{chrom_lengths}
\alias{chrom_lengths}
\title{Get Chromosome Lengths from Dataframe attributes}
\usage{
chrom_lengths(.data)
}
\arguments{
\item{.data}{Dataframe containing genomic range data.}
}
\description{
Get Chromosome Lengths from Dataframe attributes
}
