% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipboard.R
\name{clip_it}
\alias{clip_it}
\title{Copy data frame to clipboard}
\usage{
clip_it(df, col_names = TRUE, sep = "\\t", na = "", clipboard.size = 16)
}
\arguments{
\item{df}{The data frame or tibble to copy.}

\item{col_names}{Whether to include column names. Defaults to TRUE.}

\item{sep}{The separator to use between fields. Defaults to "\\t" (tab-separated).}

\item{na}{The string to use for NA values. Defaults to an empty string.}

\item{clipboard.size}{The size of the clipboard in powers of 2. Defaults to 16.}
}
\value{
None. The data is copied to the clipboard.
}
\description{
Copy data frame to clipboard
}
