% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipboard.R
\name{clipped}
\alias{clipped}
\title{Read data frame from clipboard}
\usage{
clipped(header = TRUE)
}
\arguments{
\item{header}{Whether the clipboard data includes a header. Defaults to TRUE.}
}
\value{
A tibble containing the clipboard data.
}
\description{
Read data frame from clipboard
}
