% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipboard.R
\name{clipped_vec}
\alias{clipped_vec}
\title{Read vector from clipboard}
\usage{
clipped_vec(header = FALSE)
}
\arguments{
\item{header}{Whether the clipboard data includes a header. Defaults to FALSE.}
}
\value{
A vector containing the clipboard data.
}
\description{
Read vector from clipboard
}
