% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{count_all}
\alias{count_all}
\title{Count Occurrences of Each Unique Value in All Columns}
\usage{
count_all(df)
}
\arguments{
\item{df}{A data frame for which to count occurrences of unique values in all columns.}
}
\value{
A tibble where each pair of columns represents the unique values and their counts
from a column in the input data frame. The first of each pair is the unique values,
and the second is the counts. Each pair of columns is named after the original column,
with the count columns having '_counts' appended.
}
\description{
This function applies a count operation to every column in the provided data frame. It
counts the occurrences of each unique value in each column, including \code{NA} values.
The result is a tibble with each column's unique values and their corresponding counts.
}
\examples{
\dontrun{
  data(mtcars)
  # Assuming 'mtcars' is a data frame loaded in the environment.
  count_results <- count_all(mtcars)
  print(count_results)
}
}
