% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.r
\name{create_windows}
\alias{create_windows}
\title{Bin a Dataframe of Genomic Ranges into Windows}
\usage{
create_windows(.data, width)
}
\arguments{
\item{.data}{Dataframe with columns \code{chrom}, \code{start}, \code{end}}

\item{width}{Width of window in base pairs.}

\item{chrom_lengths}{Dataframe of chromosome names and lengths in columns
\code{chrom} and \code{length}.}

\item{chroms}{Optional vector of chromosome order.}
}
\description{
Simple strand-ignoring binning procedure for genomic ranges stored in a
dataframe. The dataframe must have columns \code{chrom}, \code{start},
\code{end}.
}
