% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_manipulation.R
\name{df_map}
\alias{df_map}
\title{Apply a Function to Corresponding Columns of Two Dataframes}
\usage{
df_map(df1, df2, .f, cols = dplyr::everything(), ...)
}
\arguments{
\item{df1}{A dataframe to which the function will be applied.}

\item{df2}{Another dataframe to which the function will be applied.}

\item{.f}{A function or formula to apply to corresponding columns of df1 and df2.}

\item{cols}{A selection specification of columns to which the function will be applied.
This can be standard column names, helper functions like starts_with(),
or conditional checks like where(is.numeric). Defaults to everything().}

\item{...}{Additional arguments to pass to .f.}
}
\value{
A dataframe with the same columns as specified by cols,
each column being the result of applying .f to the corresponding
columns of df1 and df2.
}
\description{
This function applies a user-defined function (.f) to corresponding columns
of two dataframes (df1 and df2) after selecting columns based on dplyr-style
selection arguments (cols).
}
\examples{
\dontrun{
df1 <- data.frame(a = 1:3, b = 4:6)
df2 <- data.frame(a = 7:9, b = 10:12)
df_map(df1, df2, ~.x + .y)
df_map(df1, df2, ~.x * .y, cols = starts_with("b"))
}
}
