% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.r
\name{filter_.gnibble}
\alias{filter_.gnibble}
\title{Filter function for gnibble}
\usage{
\method{filter_}{gnibble}(.data, ..., .dots)
}
\arguments{
\item{.data}{A gnibble dataframe.}

\item{...}{Name-value pairs of expressions. Use \code{NULL} to drop a variable.}

\item{.dots:}{Used to work around non-standard evaluation.}
}
\description{
Filter function for gnibble
}
