% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_str_out}
\alias{filter_str_out}
\alias{filter_out}
\title{Remove rows where a column contains a string}
\usage{
filter_str_out(
  df,
  col,
  string,
  ignore_case = TRUE,
  drop.col = FALSE,
  na.rm = FALSE
)

filter_out(
  df,
  col,
  string,
  ignore_case = TRUE,
  drop.col = FALSE,
  na.rm = FALSE
)
}
\arguments{
\item{df}{A data frame or tibble.}

\item{col}{The column to filter on.}

\item{string}{The string to look for.}

\item{ignore_case}{Whether to ignore case (default is TRUE).}

\item{drop.col}{Whether to remove the column that was filtered on (default is FALSE).}

\item{na.rm}{Whether to remove NA values (default is FALSE).}
}
\value{
A filtered data frame.
}
\description{
Remove rows where a column contains a string
}
\examples{
\dontrun{
tibble(x = c("apple", "banana", "cherry")) |>
  filter_str_out(x, "app")
}
}
