% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnibble.r
\name{gnibble}
\alias{gnibble}
\title{Create a gnibble -- a genome tibble}
\usage{
gnibble(..., chrom_lengths)
}
\arguments{
\item{...}{Arguments passed to \code{tibble()}.}

\item{chrom_lengths}{Dataframe of chromosome lengths, with columns \code{chrom}, \code{lengths}.}
}
\description{
Create a gnibble -- a genome tibble
}
