% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_in_df.R
\name{is_in_df}
\alias{is_in_df}
\title{Check if a column exists in a dataframe or tibble}
\usage{
is_in_df(df, column)
}
\arguments{
\item{df}{A dataframe or tibble.}

\item{column}{A symbol representing the column name to check.}
}
\value{
A logical value: \code{TRUE} if the column exists in the dataframe or tibble, and \code{FALSE} otherwise.
}
\description{
This function determines if a specified column (passed as a symbol) exists in a given dataframe or tibble.
}
\examples{
\dontrun{
library(tibble)

df <- tibble(a = 1:5, b = 6:10)
is_in_df(df, a) # returns TRUE
is_in_df(df, z) # returns FALSE
}

}
