% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_manipulation.R
\name{listify_df}
\alias{listify_df}
\title{Split a dataframe into a list of dataframes based on unique values in a specified column.}
\usage{
listify_df(df, col, drop_col = TRUE)
}
\arguments{
\item{df}{A dataframe to split.}

\item{col}{Column name based on which to split the dataframe.}

\item{drop_col}{Indicates whether to drop the splitting column in the resultant dataframes.}
}
\value{
A named list of dataframes, each corresponding to a unique value in \code{col}.
}
\description{
Split a dataframe into a list of dataframes based on unique values in a specified column.
}
\examples{
\dontrun{
  data <- data.frame(group = c("A", "B", "A"), value = 1:3)
  list_of_dfs <- listify_df(data, group)
}

}
