% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate.R
\name{merge_if_na}
\alias{merge_if_na}
\title{Merge NA values from a target column with another column}
\usage{
merge_if_na(df, target, use_if_na, drop_col = TRUE)
}
\arguments{
\item{df}{Data frame or tibble to manipulate.}

\item{target}{The name of the target column with NA values.}

\item{use_if_na}{The name of the column to use for replacing NA values.}

\item{drop_col}{Logical indicating whether to drop the 'use_if_na' column. Defaults to TRUE.}
}
\value{
A data frame or tibble with NA values in the target column replaced.
}
\description{
This function replaces NA values in a target column with values from another specified column.
}
