% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.r
\name{mutate_.gnibble}
\alias{mutate_.gnibble}
\title{Mutate function for gnibble}
\usage{
\method{mutate_}{gnibble}(.data, ..., .dots)
}
\arguments{
\item{.data}{A gnibble dataframe.}

\item{...}{Name-value pairs of expressions. Use \code{NULL} to drop a variable.}

\item{.dots:}{Used to work around non-standard evaluation.}
}
\description{
Mutate function for gnibble
}
