% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_cell.R
\name{mutate_cell}
\alias{mutate_cell}
\title{Modify a specific cell in a data frame}
\usage{
mutate_cell(df, column, row_number, new_value)
}
\arguments{
\item{df}{A data frame that contains the cell to be modified.}

\item{column}{The name of the column (unquoted) that contains the cell to be modified.}

\item{row_number}{The row number of the cell to be modified.}

\item{new_value}{The new value to assign to the specified cell.}
}
\value{
A data frame with the specified cell modified to the new value.
The original data frame remains unchanged.
}
\description{
This function allows for the modification of a specific cell in a data frame
by specifying the column name, row number, and new value.
}
\examples{
\dontrun{
data <- data.frame(A = c(1, 2, 3), B = c("a", "b", "c"))
mutate_cell(data, A, 2, 100)
}

}
