% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_manipulation.R
\name{mutate_rowwise}
\alias{mutate_rowwise}
\title{Row-wise Mutation of Data Frame}
\usage{
mutate_rowwise(df, new_col_name, cols, fun, ...)
}
\arguments{
\item{df}{A data frame or tibble to be manipulated.}

\item{new_col_name}{A symbol indicating the name of the new column to be created.}

\item{cols}{A tidy-select expression to select columns for row-wise operation.}

\item{fun}{A function or formula to apply to each row on selected columns.}

\item{...}{Additional arguments to be passed to the function \code{fun}.}
}
\value{
A data frame with the new column added.
}
\description{
This function performs a row-wise operation on selected columns of a data
frame and creates a new column with the results.
}
\examples{
\dontrun{
  tibble(a1 = 1:5, a2 = 6:10) \%>\%
    mutate_rowwise(new_column, starts_with("a"), sum)
}
}
