% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{na_review}
\alias{na_review}
\title{Review NA Values in a Data Frame}
\usage{
na_review(df, ...)
}
\arguments{
\item{df}{A data frame object to be reviewed.}

\item{...}{use tidyselect to select columns. Defaults to all columns.}
}
\value{
A data frame with each input column's class, row count, count of NA values, and
percentage of NA values. If a column has multiple classes, only the first is listed.
}
\description{
This function provides a summary review of NA values in a data frame. It returns a data frame
that includes the class of each column, the total row count, the count and percentage of NA
values for each column. It utilizes non-standard evaluation to allow for the direct
use of unquoted column names.
}
\examples{
\dontrun{
  # Assuming 'my_data' is a data frame with columns 'a', 'b', and 'c'.
  na_review_result <- na_review(my_data, a, b, c)
  print(na_review_result)
}
}
