% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_related.R
\name{na_to_F}
\alias{na_to_F}
\title{Convert NA values to FALSE}
\usage{
na_to_F(x)
}
\arguments{
\item{x}{A vector (numeric, logical, character, etc.) where \code{NA}
values need to be replaced with \code{FALSE}.}
}
\value{
A vector of the same type as \code{x} where all \code{NA} values
have been replaced with \code{FALSE}.
}
\description{
\code{na_to_F} takes a vector and replaces all \code{NA} values with \code{FALSE},
leaving all other values unchanged. This function is useful for
handling missing data in logical vectors or conditions.
}
\examples{
\dontrun{
vec <- c(NA, 1, 0, NA, TRUE, FALSE)
na_to_F(vec)
}

}
