% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{parse_guess_all}
\alias{parse_guess_all}
\title{Reparses everything in a data frame}
\usage{
parse_guess_all(df)
}
\arguments{
\item{df}{data frame}
}
\value{
a data frame
}
\description{
Reparses everything in a data frame
}
\examples{
parse_guess_all (tibble::tibble (a = "1", b = "a", c = 5))
}
