% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peek.R
\name{peek}
\alias{peek}
\title{Peek at Transformed Object}
\usage{
peek(obj, .f = ~.x, ...)
}
\arguments{
\item{obj}{An object to be transformed.}

\item{.f}{A function to apply to the object. Default is identity function.}

\item{...}{Additional arguments to pass to \code{.f}.}
}
\value{
The original object, returned invisibly.
}
\description{
This function takes an object, applies a function to it, and prints the result, all while returning the original object.
}
\examples{
\dontrun{
library(magrittr)
mtcars \%>\% peek(~ .x \%>\% head())
}
}
