% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peek.R
\name{peek_n}
\alias{peek_n}
\title{Peek at Transformed Object with Row Limit}
\usage{
peek_n(obj, .f = ~.x, ..., n = 40)
}
\arguments{
\item{obj}{An object to be transformed.}

\item{.f}{A function to apply to the object. Default is identity function.}

\item{...}{Additional arguments to pass to \code{.f}.}

\item{n}{The maximum number of rows to print. Default is 20.}
}
\value{
The original object, returned invisibly.
}
\description{
This function takes an object, applies a function to it, and prints the result up to a specified number of rows, all while returning the original object.
}
\examples{
\dontrun{
library(magrittr)
mtcars \%>\% peek_n(~ .x \%>\% head(), n = 10)
}
}
