% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peek.R
\name{peek_v}
\alias{peek_v}
\title{Peek at Transformed Data Frame with View}
\usage{
peek_v(df, .f = ~.x, ...)
}
\arguments{
\item{df}{A data frame to be transformed.}

\item{.f}{A function to apply to the data frame. Default is identity function.}

\item{...}{Additional arguments to pass to \code{.f}.}
}
\value{
The original data frame, returned invisibly.
}
\description{
This function takes a data frame, applies a function to it, and opens the result in the RStudio viewer, all while returning the original data frame.
}
\examples{
\dontrun{
library(magrittr)
mtcars \%>\% peek_v(~ .x \%>\% head())
}
}
