% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_assign.R
\name{pipe_assign}
\alias{pipe_assign}
\title{Pipe and Assign}
\usage{
pipe_assign(obj, name, .f = ~.x, ..., envir = parent.frame(n = 2))
}
\arguments{
\item{obj}{An object to be transformed.}

\item{name}{The name of the new variable to which the result will be assigned.}

\item{.f}{A function to apply to the object. Default is identity function.}

\item{...}{Additional arguments to pass to \code{.f}.}

\item{envir}{The environment to which the new variable will be assigned. Default is the environment where the function was called.}
}
\value{
The original object.
}
\description{
This function takes an object, applies a function to it, and assigns the result to a new variable, all within a pipe.
}
\examples{
\dontrun{
library(magrittr)
mtcars \%>\% pipe_assign("mtcars_modified", ~ .x \%>\% mutate(mpg = mpg * 2))
}
}
