% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.r
\name{pos2range}
\alias{pos2range}
\title{Create Ranges from a Single SNP 'pos' Column}
\usage{
pos2range(.data)
}
\arguments{
\item{.data}{A gnibble dataframe.}
}
\description{
Create Ranges from a Single SNP 'pos' Column
}
