% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate.R
\name{quickm}
\alias{quickm}
\title{Quickly apply a function to multiple columns in a data frame}
\usage{
quickm(df, .cols, .f = ~.x, ...)
}
\arguments{
\item{df}{A data frame or tibble.}

\item{.cols}{A selection of columns where the function will be applied.}

\item{.f}{Function to be applied across the selected columns. Default is the identity function \code{~.x}.}

\item{...}{Additional arguments to be passed to the function \code{.f}.}
}
\value{
A modified data frame with the function applied to the specified columns.
}
\description{
Quickly apply a function to multiple columns in a data frame
}
\examples{
\dontrun{
library(gplyr)
library(dplyr)

df <- tibble(a = 1:3, b = 4:6, c = 7:9)
df_new <- quickm(df, .cols = c("a", "b"), .f = ~ .x * 2)
}
}
