% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{select_cols_default_all}
\alias{select_cols_default_all}
\title{Internal Function to Select Columns, with Default to All}
\usage{
select_cols_default_all(.df, ..., include_grouped = FALSE)
}
\arguments{
\item{.df}{A dataframe or tibble}

\item{...}{Columns to select, defaulting to all columns if none are specified}

\item{include_grouped}{Logical, should grouped columns be included? Defaults to FALSE.}
}
\value{
A character vector of column names
}
\description{
This internal function returns a character vector of selected column names.
}
