% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.r
\name{separate_window}
\alias{separate_window}
\title{Separate the Window Column into its Metadata columns \code{chrom},
\code{start}, \code{end}}
\usage{
separate_window(.data, remove = TRUE)
}
\arguments{
\item{.data}{Dataframe with a column \code{window} (and \code{windows}
attribute).}

\item{remove}{If \code{TRUE}, remove columns \code{chrom}, \code{start}, 
and \code{end}.}
}
\description{
Separate the Window Column into its Metadata columns \code{chrom},
\code{start}, \code{end}
}
