% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.r
\name{set_chrom_lengths}
\alias{set_chrom_lengths}
\title{Set Chromosome Lengths of a gnibble in Place}
\usage{
set_chrom_lengths(.data, value)
}
\arguments{
\item{.data}{Dataframe containing genomic range data.}

\item{value}{Dataframe of chromosome lengths (with columns
\code{chrom} and \code{length}.}
}
\description{
Set Chromosome Lengths of a gnibble in Place
}
