% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_related.R
\name{sum_rna}
\alias{sum_rna}
\title{Sum values with removal of NA values}
\usage{
sum_rna(...)
}
\arguments{
\item{...}{Numeric values or numeric vectors to be summed.}
}
\value{
The sum of all provided values with NA values removed.
}
\description{
This function sums all the provided values, automatically removing any NA values.
It is a simple wrapper around the base R \code{sum} function with \code{na.rm = TRUE} set by default.
}
\examples{
\dontrun{
sum_rna(1, 2, 3, NA, 5)  # Returns 11
sum_rna(c(1, 2, NA, 4), c(NA, 6))  # Returns 13
}

}
