% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.r
\name{summarize_window}
\alias{summarize_window}
\title{Summarize a Window}
\usage{
summarize_window(.data, ...)
}
\arguments{
\item{.data}{A gnibble with a \code{window} column.}
}
\description{
This is the equivalent to \code{dplyr}'s \code{summarize()}, except that it
carries forward relevant columns like \code{chrom}, \code{wstart},
\code{wend}, etc.
}
