% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate.R
\name{to_character}
\alias{to_character}
\title{Convert specified columns to character type}
\usage{
to_character(df, .cols = dplyr::everything())
}
\arguments{
\item{df}{A data frame or tibble.}

\item{.cols}{A selection of columns where the conversion will be applied. Defaults to all columns.}
}
\value{
A modified data frame with the specified columns converted to character type.
}
\description{
This function uses \code{quickm} to convert the specified columns to character type.
}
\examples{
\dontrun{
library(gplyr)
library(dplyr)

df <- tibble(a = 1:3, b = 4:6)
df_new <- to_character(df)
}
}
