% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate.R
\name{to_number}
\alias{to_number}
\title{Convert specified columns to numeric type}
\usage{
to_number(df, .cols = dplyr::everything())
}
\arguments{
\item{df}{A data frame or tibble.}

\item{.cols}{A selection of columns where the conversion will be applied. Defaults to all columns.}
}
\value{
A modified data frame with the specified columns converted to numeric type.
}
\description{
This function uses \code{quickm} to convert the specified columns to numeric type.
}
\examples{
\dontrun{
library(gplyr)
library(dplyr)

df <- tibble(a = c("1", "2", "3"), b = c("4", "5", "6"))
df_new <- to_number(df)
}
}
