% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloak.R
\name{uncloak}
\alias{uncloak}
\title{Reveal hidden metadata from a cloaked object}
\usage{
uncloak(x)
}
\arguments{
\item{x}{An object created with \code{cloak()}.}
}
\value{
A list of hidden metadata, or \code{NULL} if no metadata is found.
}
\description{
This function extracts the secret metadata stored via \code{cloak()}.
}
