% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.r
\name{unite_window}
\alias{unite_window}
\title{Unite Window Columns into Single Window Key Column}
\usage{
unite_window(.data, remove = TRUE)
}
\arguments{
\item{.data}{Dataframe with columns \code{chrom}, \code{start}, and 
\code{end}, (and \codes{windows} attribute).}

\item{remove}{If \code{TRUE}, remove columns \code{chrom}, \code{start}, 
and \code{end}.}
}
\description{
Unite Window Columns into Single Window Key Column
}
