% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.r
\name{windows<-}
\alias{windows<-}
\title{Set the Windows of a gnibble Object}
\usage{
windows(.data) <- value
}
\arguments{
\item{.data}{A gnibble dataframe.}

\item{value}{A gnibble of windows (with \code{chrom}, \code{start}, \code{end}).}
}
\description{
Set the Windows of a gnibble Object
}
