% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipboard.R
\name{fix_path}
\alias{fix_path}
\title{Convert Backslashes to Forward Slashes in Clipboard Paths and Quote Them}
\usage{
fix_path()
}
\value{
Invisible NULL. The function's primary effect is to alter the clipboard content.
}
\description{
This function reads a string from the clipboard, replaces all backslashes (\\)
with forward slashes (/), wraps the modified string in double quotes, and
then writes it back to the clipboard. It's intended to convert file paths
copied from the file system in Windows to a format suitable for use in R.
The function throws an error if the clipboard content is not a single string.
}
\examples{
\dontrun{
  fix_path() # Reads the clipboard, modifies its content, and updates the clipboard.
}
}
